kvutils: Extract out a trait for post-execution finalization.
